+++ CHANGELOG.md
@@ -9,10 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Marca do projeto (ícone, splash e logo)
 - Conexão com a api de vídeos: YouTube (https://www.googleapis.com/youtube/v3)
 - Layout inspirado no NetFlix
 - Adicionar o curso e disciplinas do senai
+
+## [0.2.0] - 2020-07-03
+
+### Added
+
+- Header das pages
+- Marca do projeto (ícone, splash e logo)
+
 
 ## [0.1.0] - 2020-07-01
 
@@ -21,4 +28,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Escopo do projeto
 
 [Unreleased]:
+[0.2.0]: https://github.com/rafaellevissa/appDevFlix/releases/tag/v0.2.0
 [0.1.0]: https://github.com/rafaellevissa/appDevFlix/releases/tag/v0.1.0